replace traverseStrides and mapStrides
