+++ src/Dialect/ONNX/ElementsAttr/Strides.cpp
@@ -11,19 +11,19 @@
 #include "src/Dialect/ONNX/ElementsAttr/Strides.hpp"
 
 #include "src/Dialect/ONNX/ElementsAttr/Arrays.hpp"
+#include "src/Dialect/ONNX/ElementsAttr/StridesRange.hpp"
 
 using namespace mlir;
 
 namespace onnx_mlir {
 
-size_t getStridesPosition(
-    ArrayRef<int64_t> indices, ArrayRef<int64_t> strides) {
+size_t getStridesPosition(ArrayRef<uint64_t> index, ArrayRef<int64_t> strides) {
   // Assert is commented out because this function is "on the fast path" called
   // for every element when iterating over DisposableElementsAttr values.
-  // assert(indices.size() == strides.size());
+  // assert(index.size() == strides.size());
   size_t pos = 0;
-  for (size_t axis = 0; axis < indices.size(); ++axis)
-    pos += indices[axis] * strides[axis];
+  for (size_t axis = 0; axis < index.size(); ++axis)
+    pos += index[axis] * strides[axis];
   return pos;
 }
 
@@ -158,23 +158,23 @@ SmallVector<int64_t, 4> untransposeDims(
   return unpermutedDims;
 }
 
-SmallVector<int64_t, 4> unflattenIndex(
-    ArrayRef<int64_t> shape, int64_t flatIndex) {
-  SmallVector<int64_t, 4> indices;
-  int64_t rank = shape.size();
+SmallVector<uint64_t, 4> unflattenIndex(
+    ArrayRef<int64_t> shape, uint64_t flatIndex) {
+  SmallVector<uint64_t, 4> index;
+  size_t rank = shape.size();
   if (rank > 0) {
-    indices.resize_for_overwrite(rank);
-    for (int64_t axis = rank - 1; axis >= 1; --axis) {
-      int64_t dimSize = shape[axis];
-      assert(dimSize > 0 && "cannot unflatten shape with zeros");
-      int64_t rem = flatIndex % dimSize;
+    index.resize_for_overwrite(rank);
+    for (size_t axis = rank - 1; axis >= 1; --axis) {
+      assert(shape[axis] > 0 && "cannot unflatten shape with zeros");
+      uint64_t dimSize = shape[axis];
+      uint64_t rem = flatIndex % dimSize;
       flatIndex /= dimSize;
-      indices[axis] = rem;
+      index[axis] = rem;
     }
-    assert(flatIndex < shape[0]);
-    indices[0] = flatIndex;
+    assert(static_cast<int64_t>(flatIndex) < shape[0]);
+    index[0] = flatIndex;
   }
-  return indices;
+  return index;
 }
 
 namespace {
@@ -183,9 +183,10 @@ void restrideArrayImpl(unsigned elementBytewidth, ArrayRef<int64_t> shape,
     ArrayRef<int64_t> srcStrides, ArrayRef<char> src,
     MutableArrayRef<char> dst) {
   assert(sizeof(T) == elementBytewidth && "dispatch safety check");
-  StridedArrayRef<T> stridedSrcT(castArrayRef<T>(src), srcStrides);
+  ArrayRef<T> srcT = castArrayRef<T>(src);
   MutableArrayRef<T> dstT = castMutableArrayRef<T>(dst);
-  mapStrides<T, T>(shape, dstT, stridedSrcT, [&](T elem) { return elem; });
+  for (auto &idxpos : StridesRange<1>(shape, {srcStrides}))
+    dstT[idxpos.flattenedIndex] = srcT[idxpos[0]];
 }
 } // namespace
